# prototype/reading-queue-reference/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reading Queue Testbed Public API",
    "description": "",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Workspace process is ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reading-entries": {
      "get": {
        "operationId": "listReadingEntries",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 50,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reading entry page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingEntryPageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter or cursor",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createReadingEntry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReadingEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Reading entry created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingEntryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/reading-entries/{id}/complete": {
      "post": {
        "operationId": "completeReadingEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Opaque reading entry identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reading entry completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Reading entry not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid state transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/reading-entries/{id}/reopen": {
      "post": {
        "operationId": "reopenReadingEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Opaque reading entry identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reading entry reopened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Reading entry not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid state transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateReadingEntryRequest": {
        "type": "object",
        "required": [
          "title",
          "sourceUrl"
        ],
        "properties": {
          "sourceUrl": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "traceId"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "ReadingEntryPageResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReadingEntryResponse"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ReadingEntryResponse": {
        "type": "object",
        "required": [
          "id",
          "title",
          "sourceUrl",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReadingStatusResponse"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "ReadingStatusResponse": {
        "type": "string",
        "enum": [
          "queued",
          "completed"
        ]
      }
    }
  }
}